AxisUpdata

+++ ToolFunction/Tool.cs
@@ -200,7 +200,7 @@ namespace ToolFunction
 
             return (int)(us * unit);
         }
-        public static bool CheckTimeOverSec(int tick, int time)
+        public static bool CheckTimeOverSec(long tick, int time)
         {
             var time_count = Environment.TickCount - tick;
             bool res = time_count > time * 1000;
